delete unused plan and add some tests
