--- src/artemis/experiment_plans/oav_centring_plan.py
@@ -1,271 +1,0 @@
-import bluesky.plan_stubs as bps
-import numpy as np
-from bluesky.run_engine import RunEngine
-from dodal.devices.backlight import Backlight
-from dodal.devices.oav.oav_calculations import (
-    camera_coordinates_to_xyz,
-    check_i_within_bounds,
-    extract_pixel_centre_values_from_rotation_data,
-    find_midpoint,
-    get_rotation_increment,
-    keep_inside_bounds,
-)
-from dodal.devices.oav.oav_detector import OAV
-from dodal.devices.oav.oav_errors import OAVError_WaveformAllZero
-from dodal.devices.oav.oav_parameters import OAVParameters
-from dodal.devices.smargon import Smargon
-
-from artemis.device_setup_plans.setup_oav import pre_centring_setup_oav
-from artemis.log import LOGGER, set_up_logging_handlers
-from artemis.utils.oav_utils import get_waveforms_to_image_scale
-
-# Z and Y bounds are hardcoded into GDA (we don't want to exceed them). We should look
-# at streamlining this
-_Y_LOWER_BOUND = _Z_LOWER_BOUND = -1500
-_Y_UPPER_BOUND = _Z_UPPER_BOUND = 1500
-
-# The smargon can  rotate indefinitely, so the [high/low]_limit_travel is set as 0 to
-# reflect this. Despite this, Neil would like to have omega to oscillate so we will
-# hard code limits so gridscans will switch rotation directions and |omega| will stay pretty low.
-_DESIRED_HIGH_LIMIT = 181
-
-
-def rotate_pin_and_collect_positional_data(
-    oav: OAV, smargon: Smargon, rotations: int, omega_high_limit: float
-):
-    """
-    Calculate relevant spacial values (waveforms, and pixel positions) at each rotation and save them in lists.
-
-    Args:
-        oav (OAV): The oav device to rotate and sample MXSC data from.
-        smargon (Smargon): The smargon controller device.
-        rotations (int): The number of rotations to sample.
-        omega_high_limit (float): The motor limit that shouldn't be exceeded.
-    Returns:
-        Relevant lists for each rotation, where index n corresponds to data at rotation n:
-            i_positions: the i positions of centres (x in camera coordinates)
-            j_positions: the j positions of centres (y in camera coordinates)
-            widths: the widths between the top and bottom waveforms at the centre point
-            omega_angles: the angle of the goniometer at which the measurement was taken
-            tip_i_positions: the measured i tip at a given rotation
-            tip_j_positions: the measured j tip at a given rotation
-    """
-    smargon.wait_for_connection()
-    current_omega = yield from bps.rd(smargon.omega)
-
-    # The angle to rotate by on each iteration.
-    increment = get_rotation_increment(rotations, current_omega, omega_high_limit)
-
-    # Arrays to hold positions data of the pin at each rotation,
-    # these need to be np arrays for their use in centring.
-    i_positions = np.array([], dtype=np.int32)
-    j_positions = np.array([], dtype=np.int32)
-    widths = np.array([], dtype=np.int32)
-    omega_angles = np.array([], dtype=np.int32)
-    tip_i_positions = np.array([], dtype=np.int32)
-    tip_j_positions = np.array([], dtype=np.int32)
-
-    for n in range(rotations):
-        current_omega = yield from bps.rd(smargon.omega)
-        top = np.array((yield from bps.rd(oav.mxsc.top)))
-
-        bottom = np.array((yield from bps.rd(oav.mxsc.bottom)))
-        tip_i = yield from bps.rd(oav.mxsc.tip_x)
-        tip_j = yield from bps.rd(oav.mxsc.tip_y)
-
-        for waveform in (top, bottom):
-            if np.all(waveform == 0):
-                raise OAVError_WaveformAllZero(
-                    f"Error at rotation {current_omega}, one of the waveforms is all 0"
-                )
-
-        (i, j, width) = find_midpoint(top, bottom)
-        i_positions = np.append(i_positions, i)
-        j_positions = np.append(j_positions, j)
-        widths = np.append(widths, width)
-        omega_angles = np.append(omega_angles, current_omega)
-        tip_i_positions = np.append(tip_i_positions, tip_i)
-        tip_j_positions = np.append(tip_j_positions, tip_j)
-
-        # rotate the pin to take next measurement, unless it's the last measurement
-        if n < rotations - 1:
-            yield from bps.mv(
-                smargon.omega,
-                current_omega + increment,
-            )
-
-    return (
-        i_positions,
-        j_positions,
-        widths,
-        omega_angles,
-        tip_i_positions,
-        tip_j_positions,
-    )
-
-
-def centring_plan(
-    oav: OAV,
-    parameters: OAVParameters,
-    smargon: Smargon,
-    max_run_num=3,
-    rotation_points=6,
-):
-    """
-    Attempts to find the centre of the pin on the oav by rotating and sampling elements.
-    I03 gets the number of rotation points from gda.mx.loop.centring.omega.steps which defaults to 6.
-
-    Args:
-        oav (OAV): The OAV device in use.
-        parameters (OAVParamaters): Object containing values loaded in from various parameter files in use.
-        max_run_num (int): Maximum number of times to run.
-        rotation_points (int): Test to see if the pin is widest `rotation_points` number of times on a full 180 degree rotation.
-    """
-
-    LOGGER.info("OAV Centring: Starting loop centring")
-    yield from bps.wait()
-
-    # Set relevant PVs to whatever the config dictates.
-    yield from pre_centring_setup_oav(oav, parameters)
-
-    LOGGER.info("OAV Centring: Camera set up")
-
-    # If omega  can rotate indefinitely (indicated by high_limit_travel=0), we set the hard coded limit.
-    omega_high_limit = yield from bps.rd(smargon.omega.high_limit_travel)
-    if not omega_high_limit:
-        omega_high_limit = _DESIRED_HIGH_LIMIT
-
-    # The image resolution may not correspond to the (1024, 768) of the waveform, then we have to scale
-    # waveform pixels to get the camera pixels.
-    i_scale, j_scale = yield from get_waveforms_to_image_scale(oav)
-
-    # array for holding the current xyz position of the motor.
-    motor_xyz = np.array(
-        [
-            (yield from bps.rd(smargon.x)),
-            (yield from bps.rd(smargon.y)),
-            (yield from bps.rd(smargon.z)),
-        ],
-        dtype=np.float64,
-    )
-
-    LOGGER.info(f"OAV Centring: Starting xyz, {motor_xyz}")
-
-    # We attempt to find the centre `max_run_num` times...
-    for run_num in range(max_run_num):
-        # Spin the goniometer and capture data from the camera at each rotation_point.
-        (
-            i_positions,
-            j_positions,
-            widths,
-            omega_angles,
-            tip_i_positions,
-            tip_j_positions,
-        ) = yield from rotate_pin_and_collect_positional_data(
-            oav, smargon, rotation_points, omega_high_limit
-        )
-
-        LOGGER.info(
-            f"Run {run_num}, mid_points: {(i_positions, j_positions)}, widths {widths}, angles {omega_angles}, tips {(tip_i_positions, tip_j_positions)}"
-        )
-
-        # Filters the data captured at rotation and formats it in terms of i, j, k and angles.
-        # (i_pixels,j_pixels) correspond to the (x,y) midpoint at the widest rotation, in the camera coordinate system,
-        # k_pixels correspond to the distance between the midpoint and the tip of the camera at the angle orthogonal to the
-        # widest rotation.
-        (
-            i_pixels,
-            j_pixels,
-            k_pixels,
-            best_omega_angle,
-            best_omega_angle_orthogonal,
-        ) = extract_pixel_centre_values_from_rotation_data(
-            i_positions, j_positions, widths, omega_angles
-        )
-
-        LOGGER.info(f"Run {run_num} centre in pixels {(i_pixels, j_pixels, k_pixels)}")
-        LOGGER.info(
-            f"Run {run_num} best angles {(best_omega_angle, best_omega_angle_orthogonal)}"
-        )
-
-        # Adjust waveform values to match the camera pixels.
-
-        i_pixels *= i_scale
-        j_pixels *= j_scale
-        k_pixels *= j_scale
-        LOGGER.info(
-            f"Run {run_num} centre in pixels after scaling {(i_pixels, j_pixels, k_pixels)}"
-        )
-
-        # Adjust i_pixels if it is too far away from the pin.
-        tip_i = np.median(tip_i_positions)
-
-        i_pixels = check_i_within_bounds(
-            parameters.max_tip_distance_pixels, tip_i, i_pixels
-        )
-        # Get the beam distance from the centre (in pixels).
-        (
-            beam_distance_i_pixels,
-            beam_distance_j_pixels,
-        ) = parameters.calculate_beam_distance(i_pixels, j_pixels)
-
-        # Add the beam distance to the current motor position (adjusting for the changes in coordinate system
-        # and the from the angle).
-        motor_xyz += camera_coordinates_to_xyz(
-            beam_distance_i_pixels,
-            beam_distance_j_pixels,
-            best_omega_angle,
-            parameters.micronsPerXPixel,
-            parameters.micronsPerYPixel,
-        )
-
-        LOGGER.info(f"Run {run_num} move for x, y {motor_xyz}")
-
-        if run_num == max_run_num - 1:
-            # If it's the last run we adjust the z value of the motors.
-            beam_distance_k_pixels = parameters.calculate_beam_distance(
-                i_pixels, k_pixels
-            )[1]
-
-            motor_xyz += camera_coordinates_to_xyz(
-                0,
-                beam_distance_k_pixels,
-                best_omega_angle_orthogonal,
-                parameters.micronsPerXPixel,
-                parameters.micronsPerYPixel,
-            )
-
-        # If the x value exceeds the stub offsets, reset it to the stub offsets
-        motor_xyz[1] = keep_inside_bounds(motor_xyz[1], _Y_LOWER_BOUND, _Y_UPPER_BOUND)
-        motor_xyz[2] = keep_inside_bounds(motor_xyz[2], _Z_LOWER_BOUND, _Z_UPPER_BOUND)
-
-        run_num += 1
-        LOGGER.info(f"Run {run_num} move for x, y, z: {motor_xyz}")
-
-        yield from bps.mv(
-            smargon.x, motor_xyz[0], smargon.y, motor_xyz[1], smargon.z, motor_xyz[2]
-        )
-
-    # We've moved to the best x,y,z already. Now rotate to the widest pin angle.
-    yield from bps.mv(smargon.omega, best_omega_angle)
-
-    yield from bps.sleep(1)
-    LOGGER.info("Finished loop centring")
-
-
-if __name__ == "__main__":
-    beamline = "BL03I"
-    set_up_logging_handlers("INFO")
-    oav = OAV(name="oav", prefix=beamline)
-
-    smargon: Smargon = Smargon(name="smargon", prefix=beamline)
-    backlight: Backlight = Backlight(name="backlight", prefix=beamline)
-    parameters = OAVParameters(
-        "src/artemis/unit_tests/test_OAVCentring.json",
-        "src/artemis/unit_tests/test_jCameraManZoomLevels.xml",
-        "src/artemis/unit_tests/test_display.configuration",
-    )
-    oav.wait_for_connection()
-    smargon.wait_for_connection()
-    RE = RunEngine()
-    RE(centring_plan(oav, parameters, smargon, backlight))
