+++ src/artemis/experiment_plans/tests/test_full_grid_scan_plan.py
@@ -1,0 +1,38 @@
+from unittest.mock import patch
+
+import pytest
+from bluesky.run_engine import RunEngine
+from dodal.i03 import detector_motion
+
+from artemis.experiment_plans.full_grid_scan import (
+    create_devices,
+    get_plan,
+    wait_for_det_to_finish_moving,
+)
+
+
+def test_create_devices():
+    with (
+        patch("artemis.experiment_plans.full_grid_scan.i03") as i03,
+        patch(
+            "artemis.experiment_plans.full_grid_scan.fgs_create_devices"
+        ) as fgs_create_devices,
+        patch(
+            "artemis.experiment_plans.full_grid_scan.oav_create_devices"
+        ) as oav_create_devices,
+    ):
+        create_devices()
+        fgs_create_devices.assert_called()
+        oav_create_devices.assert_called()
+        i03.detector_motion.return_value.wait_for_connection.assert_called()
+        i03.backlight.return_value.wait_for_connection.assert_called()
+
+
+def test_wait_for_detector():
+    d_m = detector_motion(fake_with_ophyd_sim=True)
+    RE = RunEngine({})
+    with pytest.raises(TimeoutError):
+        RE(wait_for_det_to_finish_moving(d_m, 0.2))
+    d_m.shutter.sim_put(1)
+    d_m.z.motor_done_move.sim_put(1)
+    RE(wait_for_det_to_finish_moving(d_m, 0.5))
